add legacy stream
